# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Desastres",
        "description": "Esta es la documentación de la API del proyecto Desastres.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ubications": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Listar todas las ubicaciones",
                "operationId": "90d893d83a5d48b3d7c2183b55a0774e",
                "responses": {
                    "200": {
                        "description": "Lista de ubicaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ubication"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Crear una nueva ubicación",
                "operationId": "966ad334d94d9bf6c4758cb9408a934f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parque Central"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un hermoso parque en el centro de la ciudad."
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.67
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -103.35
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Principal #123"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Ciudad de México"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "CDMX"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "México"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "01000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ubicación creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubication"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ubications/{id}": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Mostrar una ubicación específica",
                "operationId": "204462025e1693c1170d61bb22f6561c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la ubicación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubication"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Actualizar una ubicación",
                "operationId": "c2e0060aa1f8bc62dd700c181c25798a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parque Central Actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un hermoso parque actualizado en el centro de la ciudad."
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.67
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -103.35
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Principal #123"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Ciudad de México"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "CDMX"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "México"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "01000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ubicación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubication"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Eliminar una ubicación",
                "operationId": "61815bfe1c2d8bde0ab71a08e23daf45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ubicación eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha eliminado la ubicación correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ubication": {
                "required": [
                    "name",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Parque Central"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un hermoso parque en el centro de la ciudad."
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 20.67
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -103.35
                    },
                    "address": {
                        "type": "string",
                        "example": "Calle Principal #123"
                    },
                    "city": {
                        "type": "string",
                        "example": "Ciudad de México"
                    },
                    "state": {
                        "type": "string",
                        "example": "CDMX"
                    },
                    "country": {
                        "type": "string",
                        "example": "México"
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "01000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token en el formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}